+++ src/audio/index.ts
@@ -8,9 +8,11 @@ export { Instrument, linearToDecibels }
 
 initCore();
 
-class AudioError extends MusicError {
+export class AudioError extends MusicError {
     constructor(message: string) {
         super(MusicErrorType.Audio, message);
+        Object.setPrototypeOf(this, new.target.prototype); // Fix prototype chain
+        this.name = new.target.name;
     }
 }