+++ src/core/error.ts
@@ -3,14 +3,15 @@
 export enum MusicErrorType {
     Unknown,
     InvalidArg,
+    Score,
+    Audio,
     Note,
+    NoteDuration,
     Scale,
     KeySignature,
     TimeSignature,
     /** @deprecated Typo, will be removed. */
     Timesignature = MusicErrorType.TimeSignature,
-    Score,
-    Audio
 }
 
 function isType(type: unknown): type is MusicErrorType {
@@ -18,7 +19,7 @@ function isType(type: unknown): type is MusicErrorType {
 }
 
 function formatType(type: MusicErrorType | undefined): string {
-    return type === undefined ? "" : `[${MusicErrorType[type]}] `;
+    return type === undefined ? "" : String(MusicErrorType[type]);
 }
 
 /** Music error class. */
@@ -42,10 +43,17 @@ export class MusicError extends Error {
             ? [MusicErrorType.Unknown, args[0]]
             : args as [MusicErrorType, string];
 
-        super(formatType(type) + msg);
+        super(`${formatType(type)} ${msg}`);
         Object.setPrototypeOf(this, new.target.prototype); // Fix prototype chain
-
-        this.name = "MusicError";
+        this.name = new.target.name;
         this.type = type;
+    }
+}
+
+export class InvalidArgError extends MusicError {
+    constructor(message: string) {
+        super(MusicErrorType.InvalidArg, message);
+        Object.setPrototypeOf(this, new.target.prototype); // Fix prototype chain
+        this.name = new.target.name;
     }
 }