+++ src/theory/guitar.ts
@@ -1,14 +1,8 @@
 import { Assert, LRUCache } from "@tspro/ts-utils-lib";
 import TuningData from "./assets/tunings.json";
 import { Note } from "./note";
-import { MusicError, MusicErrorType } from "web-music-score/core";
+import { InvalidArgError } from "web-music-score/core";
 import { getClosestString } from "shared-src";
-
-class TuningError extends MusicError {
-    constructor(message: string) {
-        super(MusicErrorType.InvalidArg, message);
-    }
-}
 
 /*
  * | string | stringId | freq    | fretboard (RH) | on tab        |
@@ -26,7 +20,7 @@ export const DefaultTuningName: string = TuningNameList[0];
 
 export function validateTuningName(tuningName: string): string {
     if (typeof tuningName !== "string" || TuningNameList.indexOf(tuningName) < 0) {
-        throw new TuningError( `Invalid tuning name "${tuningName}".`);
+        throw new InvalidArgError(`Invalid tuning name "${tuningName}".`);
     }
     else {
         return tuningName;
@@ -43,9 +37,9 @@ export function resolveTuningName(tuningName: unknown): string {
     const closest = getClosestString(tuningNameStr, TuningNameList);
 
     if (closest)
-        throw new TuningError( `Invalid tuning name "${tuningName}". Did you mean "${closest}"?`);
+        throw new InvalidArgError(`Invalid tuning name "${tuningName}". Did you mean "${closest}"?`);
 
-    throw new TuningError( `Invalid tuning name "${tuningName}".`);
+    throw new InvalidArgError(`Invalid tuning name "${tuningName}".`);
 }
 
 const TuningStringsCache = new LRUCache<string, ReadonlyArray<Note>>(100);
@@ -62,7 +56,7 @@ export function getTuningStrings(tuningName: string): ReadonlyArray<Note> {
         let tuningData = TuningData.list.find(data => data.name === tuningName);
 
         if (!tuningData) {
-            throw new TuningError( `Invalid tuning name "${tuningName}".`);
+            throw new InvalidArgError(`Invalid tuning name "${tuningName}".`);
         }
 
         tuningStrings = tuningData.strings.slice().reverse().map(noteName => Note.getNote(noteName));