+++ src/theory/key-signature.ts
@@ -3,9 +3,11 @@ import { Accidental, Note } from "./note";
 import { getScale, ScaleType } from "./scale";
 import { MusicError, MusicErrorType } from "web-music-score/core";
 
-class KeySignatureError extends MusicError {
+export class KeySignatureError extends MusicError {
     constructor(message: string) {
         super(MusicErrorType.KeySignature, message);
+        Object.setPrototypeOf(this, new.target.prototype); // Fix prototype chain
+        this.name = new.target.name;
     }
 }