+++ src/theory/note.ts
@@ -2,9 +2,11 @@ import { Guard, UniMap } from "@tspro/ts-utils-lib";
 import { PitchNotation, SymbolSet } from "./types";
 import { MusicError, MusicErrorType } from "web-music-score/core";
 
-class NoteError extends MusicError {
+export class NoteError extends MusicError {
     constructor(message: string) {
         super(MusicErrorType.Note, message);
+        Object.setPrototypeOf(this, new.target.prototype); // Fix prototype chain
+        this.name = new.target.name;
     }
 }